Updated to 1.0.3
Separated the promise methods from the normal ones and added a method for fetching offline players profiles.

Also changed the behavior of timeout to have a default value of 3 instead of a huge number, which means that if you want to wait a lot for a profile or data, you have to specify how long to wait for or it can return nil.